flask/requirements.txt
aniso8601==7.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-Login==0.4.1
flask-redis==0.4.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
httpie==2.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kafka-python==2.0.1
kazoo==2.5.0
MarkupSafe==1.1.1
psycopg2==2.8.4
Pygments==2.5.2
pykafka==2.8.0
pytz==2019.2
redis==3.4.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.13
tabulate==0.8.6
urllib3==1.25.8
Werkzeug==0.15.5
